Complete build steps add demands list on admin and demand crud operations

+++ fmarket/src/main/java/ro/fmarket/model/demand/DemandDao.java
@@ -75,7 +75,7 @@ public class DemandDao extends BaseDao<Demand> {
 
 	public int getCriteriaTotalCount(Criteria criteria) {
 		criteria.setProjection(Projections.rowCount());
-		return (int) criteria.uniqueResult();
+		return (int)(long)criteria.uniqueResult();
 	}
 
 	@SuppressWarnings("unchecked")
